Add basic commands for tracking and analysing
Create the basic project structure. Add two commands, one for tracking
expenses, one for listing all expenses for a specific month. Add a basic
database schema to write expenses into and read from, and a config
dataclass that can be read from a yml file. The database scheme is very
limited still, more will be added later. The current state is only meant
as skeleton for functionality to come.